refactor(react): share modal utilities
